Fixed calloc bug, Made Bitwise faster.
I was using calloc to allocate memory, but i was expecting the memory to
be random garbage... it turns out that i didn't read what calloc does...
which is initializes said memory, thus preventing me from getting the
garbage that I had wanted. I have since changed this to malloc to get
real responses. Although I am still not sure if it is working the way I
want.

I also updated the Bitwise function. I noticed a redundancy in that I
was checking for '0', then returning '0' or '0' by use of a trinary.
Which is just redundant seeing that I was checking for 0, i should just
return that number. so I did. I also converted away from using the gcc
specific '0b' notation for binary, and followed the standard hex
notation.